# perf(vite): communicate with vite-node via internal socket