--- a/packages/vite/src/runtime/client.manifest.mjs
+++ b/packages/vite/src/runtime/client.manifest.mjs
@@ -1,4 +1,4 @@
 // @ts-check
 import { viteNodeFetch } from './vite-node-shared.mjs'
 
-export default () => viteNodeFetch('/manifest')
+export default () => viteNodeFetch.getManifest()