--- a/packages/vite/src/runtime/vite-node-types.d.ts
+++ b/packages/vite/src/runtime/vite-node-types.d.ts
@@ -1,0 +1,4 @@
+/**
+ * Shared type definitions for Vite Node IPC communication
+ * Used by both server-side (vite-node.ts) and client-side (vite-node-shared.mjs) code
+ */