--- a/packages/vite/src/runtime/vite-node-types.d.ts
+++ b/packages/vite/src/runtime/vite-node-types.d.ts
@@ -1,4 +1,0 @@
-/**
- * Shared type definitions for Vite Node IPC communication
- * Used by both server-side (vite-node.ts) and client-side (vite-node-shared.mjs) code
- */