--- a/packages/vite/src/runtime/vite-node.mjs
+++ b/packages/vite/src/runtime/vite-node.mjs
@@ -20,7 +20,7 @@ export default async (ssrContext) => {
   import.meta.server = true
 
   // Invalidate cache for files changed since last rendering
-  const invalidates = await viteNodeFetch('/invalidates')
+  const invalidates = await viteNodeFetch.getInvalidates()
   const updates = runner.moduleCache.invalidateDepTree(invalidates)
 
   // Execute SSR bundle on demand
@@ -40,11 +40,11 @@ function createRunner () {
     root: viteNodeOptions.root, // Equals to Nuxt `srcDir`
     base: viteNodeOptions.base,
     async resolveId (id, importer) {
-      return await viteNodeFetch('/resolve/' + encodeURIComponent(id) + (importer ? '?importer=' + encodeURIComponent(importer) : '')) ?? undefined
+      return await viteNodeFetch.resolveId(id, importer)
     },
     async fetchModule (id) {
       id = id.replace(/\/\//g, '/') // TODO: fix in vite-node
-      return await viteNodeFetch('/module/' + encodeURI(id)).catch((err) => {
+      return await viteNodeFetch.fetchModule(id).catch((err) => {
         const errorData = err?.data?.data
         if (!errorData) {
           throw err