### 🔗 Linked issue

<!-- Please ensure there is an open issue and mention its number. For example, "resolves #123" -->

### 📚 Description

this migrates the vite-node server transport (from nitro worker-thread <> vite plugin) to use a raw socket rather than network communication.

the potential benefit here is mostly on Windows, where the networking layer can be very slow indeed.

socket naming is taken partially from nitro (see [getSocketAddress](https://github.com/nitrojs/nitro/blob/4971b434db0461371c54425da68a4ef2f7d4256d/src/presets/_nitro/runtime/nitro-dev.ts#L119-L134)) which saw some iteration in a previous nitro release to ensure that things worked well in dev server.

in addition, this PR migrates socket responses to be type-safe.

### 🚧 TODO

- [x] vite-node error handling

### 🏁 Benchmarks

I ran a very simple benchmark test on three platforms with a bare minimum Nuxt project, and here were the results

#### Concurrent Requests (10 concurrent, 5 iterations)

| Platform | Metric | Without PR | With PR | Change |
|----------|--------|------------|---------|---------|
| **Windows** | Average | 162.5ms | 151.43ms | **-6.8%** ✅ |
| | Median | 149.18ms | 94.5ms | **-36.7%** ✅ |
| **macOS** | Average | 22.19ms | 17.99ms | **-18.9%** ✅ |
| | Median | 19.6ms | 16.8ms | **-14.3%** ✅ |
| **Linux** | Average | 38.15ms | 42.14ms | **+10.5%** ❌ |
| | Median | 39.22ms | 45.49ms | **+16.0%** ❌ |

#### Single Requests (50 sequential)

| Platform | Metric | Without PR | With PR | Change |
|----------|--------|------------|---------|---------|
| **Windows** | Average | 21.08ms | 16.95ms | **-19.6%** ✅ |
| | Median | 20.32ms | 15.35ms | **-24.5%** ✅ |
| **macOS** | Average | 3.55ms | 3.84ms | **+8.2%** ❌ |
| | Median | 3.47ms | 3.38ms | **-2.6%** ✅ |
| **Linux** | Average | 7.83ms | 8.21ms | **+4.9%** ❌ |
| | Median | 7.06ms | 7ms | **-0.8%** ✅ |

I noticed a fair amount of variability when performing the test, but in general the Docker/Linux setup saw a regression (I've bailed on using abstract sockets over potential perf cost on docker.)